+++ Tests/TestProjectGenerator/Reqnroll.TestProjectGenerator/Factories/BindingsGenerator/FSharpBindingsGenerator.cs
@@ -111,6 +111,7 @@ type internal Log =
             string hookType,
             string name,
             string code = "",
+            bool? asyncHook = null,
             int? order = null,
             IList<string> hookTypeAttributeTags = null,
             IList<string> methodScopeAttributeTags = null,