+++ CHANGELOG.md
@@ -4,7 +4,11 @@
 
 ## Bug fixes:
 
-*Contributors of this release (in alphabetical order):* 
+* Fix: xUnit async `[AfterTestRun]` hook might not execute fully (#530)
+* Fix: Scenario, feature and test run finished event is not published when the related "after" hook fails (#560)
+* Fix: Inconsistent hook execution (duble execution, before/after hook skipped, infrastructure errors) when before or after hooks fail (#526)
+
+*Contributors of this release (in alphabetical order): @clrudolphi, @gasparnagy* 
 
 # v2.4.0 - 2025-03-06